remove new-time build flag
This caused problems building with stackage and ghc 7.6, since cabal
assumes the flag means it wants a newer time. Since time is bundled with
ghc, stackage doesn't pin it.

Since this flag was only there to avoid a dep on old-locale, which is
currently bundled with ghc, let's remove the flag.