add \usr\bin to path
This is where these commands are installed by the new "git for windows".
Kept the old paths too, so msysgit will still work.

This was enough to get ssh and ssh-keygen working when run at the DOS
prompt with the new git for windows installed. However, for some reason,
rsync still fails to find ssh in path when that version of git is used.